--- additional_scripts/check_of_procedure.py
@@ -1,8 +1,0 @@
-# Verify the existence of the requested procedure
-def check_of_procedure(procedure_name):
-    return procedure_name in [
-        "transcribe",
-        "reverse",
-        "complement",
-        "reverse_complement",
-    ]
