--- additional_scripts/reverse.py
@@ -1,3 +1,0 @@
-# Return the expanded sequence
-def reverse(chain):
-    return chain[::-1]
