--- additional_scripts/reverse_complement.py
@@ -1,3 +1,0 @@
-# Return the reverse complementary sequence
-def reverse_complement(chain):
-    return complement(reverse(chain))
